librtasevent: 'format' function attribute for printf-style functions
Clang (but not GCC) says:

librtasevent_src/print_rtas_event.c:146:61: error: format string is
not a string literal [-Werror,-Wformat-nonliteral]
    offset += vsnprintf(buf + offset, sizeof(buf) - offset, fmt, ap);
                                                            ^~~
librtasevent_src/print_rtas_event.c:314:52: error: format string is
not a string literal [-Werror,-Wformat-nonliteral]
    tmpbuf_len = vsnprintf(tmpbuf, sizeof(tmpbuf), fmt, ap);

Adding a 'format' function attribute to print_scn_title() and
rtas_print() yields the following warnings:

*** fixme ***

Fix the fallout.

Notably, this found longstanding bugs in print_re_post_scn(), which
passes pointers where integers are specified by format strings. These
sites have been fixed to print the contents of the structure fields
they were intended to report.

See discussion of -Wformat-nonliteral at
https://clang.llvm.org/docs/AttributeReference.html#format

+++ librtasevent_src/rtas_epow.c
@@ -99,7 +99,7 @@ print_v4_epow(struct scn_header *shdr, int verbosity)
     int len = 0;
 
     len += print_scn_title("EPOW Warning");
-    len += rtas_print(PRNT_FMT_R, "EPOW Sensor Value:", epow->sensor_value); 
+    len += rtas_print(PRNT_FMT_R, "EPOW Sensor Value:", (uint32_t)epow->sensor_value); 
 
     if (version >= 3) {
         if (epow->sensor) {
@@ -108,7 +108,7 @@ print_v4_epow(struct scn_header *shdr, int verbosity)
                               "Sensor Token:", epow->sensor_token, 
                               "Sensor Index:", epow->sensor_index);
             len += rtas_print(PRNT_FMT_2,
-                              "Sensor Value:", epow->sensor_value,
+                              "Sensor Value:", (uint32_t)epow->sensor_value,
                               "Sensor Status:", epow->sensor_status);
         }
 
@@ -161,8 +161,8 @@ print_v6_epow(struct scn_header *shdr, int verbosity)
 
     len += print_v6_hdr("EPOW Warning", &epow->v6hdr, verbosity);
 
-    len += rtas_print(PRNT_FMT_2, "Sensor Value:", epow->sensor_value,
-                      "Action Code:", epow->action_code);
+    len += rtas_print(PRNT_FMT_2, "Sensor Value:", (uint32_t)epow->sensor_value,
+                      "Action Code:", (uint32_t)epow->action_code);
     len += rtas_print(PRNT_FMT_R, "EPOW Event:", epow->event_modifier);
 
     switch (epow->event_modifier) {
